Improved watch provider settings UX
