+++ Shared/View/Settings/WatchProviderSettings.swift
@@ -6,12 +6,13 @@
 //
 
 import SwiftUI
+import NukeUI
 
 struct WatchProviderSettings: View {
     @StateObject private var store = SettingsStore.shared
-#if os(macOS)
-    @State private var showWatchProvidersSelector = false
-#endif
+    @StateObject private var settings = SettingsStore.shared
+    @State private var providers = [WatchProviderContent]()
+    @State private var isLoading = true
     var body: some View {
         Form {
             Section {
@@ -44,19 +45,23 @@ struct WatchProviderSettings: View {
 					Text("Watch Providers")
 					Text("See in what platforms the content is available on.")
                 }
-#if os(iOS) || os(visionOS)
-                NavigationLink("Streaming Services", destination: WatchProviderSelectorSetting())
-#elseif os(macOS)
-                Button("Streaming Services") {
-                    showWatchProvidersSelector.toggle()
+                Toggle(isOn: $settings.isSelectedWatchProviderEnabled) {
+                    Text("Preferred Streaming Services")
+                    Text("Only show 'Watch Provider' if the title is available on one of your preferred services")
                 }
-                .sheet(isPresented: $showWatchProvidersSelector) {
-                    NavigationStack {
-                        WatchProviderSelectorSetting(showView: $showWatchProvidersSelector)
+                if settings.isSelectedWatchProviderEnabled {
+                    Section {
+                        if providers.isEmpty, !isLoading {
+                            SimpleUnavailableView()
+                        } else if providers.isEmpty, isLoading {
+                            ProgressView()
+                        } else {
+                            List(providers, id: \.itemID) { item in
+                                WatchProviderItemSelector(item: item)
+                            }
+                        }
                     }
-                    .frame(width: 400, height: 500, alignment: .center)
                 }
-#endif
                 
             }
 #endif
@@ -71,6 +76,10 @@ struct WatchProviderSettings: View {
 #elseif os(macOS)
         .formStyle(.grouped)
 #endif
+        .onChange(of: settings.isSelectedWatchProviderEnabled) { _ in
+            checkStatus()
+        }
+        .onAppear(perform: load)
     }
     
 #if os(iOS)
@@ -90,4 +99,86 @@ struct WatchProviderSettings: View {
 
 #Preview {
     WatchProviderSettings()
+}
+
+extension WatchProviderSettings {
+    private func load() {
+        if !settings.isWatchProviderEnabled { return }
+        Task {
+            if !isLoading { return }
+            if providers.isEmpty {
+                do {
+                    let network = NetworkService.shared
+                    let providers = try await network.fetchWatchProviderServices(for: .movie, region: SettingsStore.shared.watchRegion.rawValue)
+                    var result = Set<WatchProviderContent>()
+                    for item in providers.results {
+                        if !result.contains(where: { $0.itemId == item.itemId }) {
+                            result.insert(item)
+                        }
+                    }
+                    self.providers.append(contentsOf: result.sorted { $0.providerTitle < $1.providerTitle})
+                    withAnimation { isLoading = false }
+                } catch {
+                    if Task.isCancelled { return }
+                    CronicaTelemetry.shared.handleMessage(error.localizedDescription,
+                                                          for: "WatchProviderSelectorSetting.load.failed")
+                }
+            }
+        }
+    }
+    
+    private func checkStatus() {
+        if settings.isSelectedWatchProviderEnabled {
+            load()
+        } else {
+            settings.selectedWatchProviders = ""
+        }
+    }
+}
+
+private struct WatchProviderItemSelector: View {
+    let item: WatchProviderContent
+    @StateObject private var settings = SettingsStore.shared
+    @State private var isSelected = false
+    var body: some View {
+        HStack {
+            Image(systemName: isSelected ? "checkmark.circle.fill" : "circle")
+                .foregroundColor(isSelected ? SettingsStore.shared.appTheme.color : nil)
+                .fontWeight(.semibold)
+                .padding(.trailing)
+            LazyImage(url: item.providerImage) { state in
+                if let image = state.image {
+                    image
+                        .resizable()
+                        .aspectRatio(contentMode: .fit)
+                } else {
+                    Rectangle().fill(.gray.gradient)
+                }
+            }
+                .frame(width: 40, height: 40, alignment: .center)
+                .clipShape(RoundedRectangle(cornerRadius: 8, style: .continuous))
+                .accessibilityHidden(true)
+            Text(item.providerTitle)
+        }
+        .onTapGesture {
+            if settings.selectedWatchProviders.contains(item.itemID) {
+                let selected = settings.selectedWatchProviders.replacingOccurrences(of: item.itemID, with: "")
+                settings.selectedWatchProviders = selected
+            } else {
+                settings.selectedWatchProviders.append(item.itemID)
+            }
+            HapticManager.shared.selectionHaptic()
+        }
+        .task(id: settings.selectedWatchProviders) {
+            if settings.selectedWatchProviders.contains(item.itemID)  {
+                if !isSelected {
+                    withAnimation { isSelected = true }
+                }
+            } else {
+                if isSelected {
+                    withAnimation { isSelected = false }
+                }
+            }
+        }
+    }
 }
